♻️ refactor: 패키지 경로를 domain.member로 변경
