+++ src/main/java/com/jajaja/domain/auth/service/CustomOAuth2UserService.java
@@ -2,15 +2,15 @@ package com.jajaja.domain.auth.service;
 
 import com.jajaja.domain.cart.entity.Cart;
 import com.jajaja.domain.cart.repository.CartRepository;
-import com.jajaja.domain.user.converter.MemberConverter;
+import com.jajaja.domain.member.converter.MemberConverter;
 import com.jajaja.domain.auth.dto.UserDto;
 import com.jajaja.domain.auth.dto.CustomOAuth2User;
 import com.jajaja.domain.auth.dto.GoogleResponseDto;
 import com.jajaja.domain.auth.dto.KakaoResponseDto;
 import com.jajaja.domain.auth.dto.OAuth2ResponseDto;
-import com.jajaja.domain.user.entity.Member;
-import com.jajaja.domain.user.entity.enums.OauthType;
-import com.jajaja.domain.user.repository.MemberRepository;
+import com.jajaja.domain.member.entity.Member;
+import com.jajaja.domain.member.entity.enums.OauthType;
+import com.jajaja.domain.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
