+++ src/main/java/com/jajaja/domain/cart/entity/Cart.java
@@ -1,7 +1,7 @@
 package com.jajaja.domain.cart.entity;
 
 import com.jajaja.domain.coupon.entity.Coupon;
-import com.jajaja.domain.user.entity.Member;
+import com.jajaja.domain.member.entity.Member;
 import com.jajaja.global.apiPayload.code.status.ErrorStatus;
 import com.jajaja.global.apiPayload.exception.custom.BadRequestException;
 import com.jajaja.global.common.domain.BaseEntity;
