+++ src/main/java/com/jajaja/domain/delivery/entity/Delivery.java
@@ -1,7 +1,7 @@
 package com.jajaja.domain.delivery.entity;
 
 import com.jajaja.domain.order.entity.Order;
-import com.jajaja.domain.user.entity.Member;
+import com.jajaja.domain.member.entity.Member;
 import com.jajaja.global.common.domain.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
