+++ src/main/java/com/jajaja/domain/member/entity/enums/OauthType.java
@@ -1,0 +1,5 @@
+package com.jajaja.domain.member.entity.enums;
+
+public enum OauthType {
+    KAKAO, GOOGLE
+}
