+++ src/main/java/com/jajaja/domain/product/service/ProductQueryServiceImpl.java
@@ -19,10 +19,10 @@ import com.jajaja.domain.review.repository.ReviewRepository;
 import com.jajaja.domain.team.dto.response.TeamResponseDto;
 import com.jajaja.domain.team.entity.Team;
 import com.jajaja.domain.team.repository.TeamRepository;
-import com.jajaja.domain.user.entity.Member;
-import com.jajaja.domain.user.entity.MemberBusinessCategory;
-import com.jajaja.domain.user.repository.MemberBusinessCategoryRepository;
-import com.jajaja.domain.user.repository.MemberRepository;
+import com.jajaja.domain.member.entity.Member;
+import com.jajaja.domain.member.entity.MemberBusinessCategory;
+import com.jajaja.domain.member.repository.MemberBusinessCategoryRepository;
+import com.jajaja.domain.member.repository.MemberRepository;
 import com.jajaja.global.apiPayload.PageResponse;
 import com.jajaja.global.apiPayload.code.status.ErrorStatus;
 import com.jajaja.global.apiPayload.exception.custom.BadRequestException;
