+++ src/main/java/com/jajaja/domain/review/entity/Review.java
@@ -2,7 +2,7 @@ package com.jajaja.domain.review.entity;
 
 import com.jajaja.domain.product.entity.Product;
 import com.jajaja.domain.product.entity.ProductOption;
-import com.jajaja.domain.user.entity.Member;
+import com.jajaja.domain.member.entity.Member;
 import com.jajaja.global.common.domain.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
