+++ src/main/java/com/jajaja/domain/team/entity/Team.java
@@ -3,7 +3,7 @@ package com.jajaja.domain.team.entity;
 import com.jajaja.domain.order.entity.Order;
 import com.jajaja.domain.product.entity.Product;
 import com.jajaja.domain.team.entity.enums.TeamStatus;
-import com.jajaja.domain.user.entity.Member;
+import com.jajaja.domain.member.entity.Member;
 import com.jajaja.global.common.domain.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
