+++ src/main/java/com/jajaja/domain/team/service/TeamCommandServiceImpl.java
@@ -8,8 +8,8 @@ import com.jajaja.domain.team.dto.response.TeamCreateResponseDto;
 import com.jajaja.domain.team.entity.Team;
 import com.jajaja.domain.team.entity.enums.TeamStatus;
 import com.jajaja.domain.team.repository.TeamRepository;
-import com.jajaja.domain.user.entity.Member;
-import com.jajaja.domain.user.repository.MemberRepository;
+import com.jajaja.domain.member.entity.Member;
+import com.jajaja.domain.member.repository.MemberRepository;
 import com.jajaja.global.apiPayload.code.status.ErrorStatus;
 import com.jajaja.global.apiPayload.exception.custom.BadRequestException;
 import lombok.RequiredArgsConstructor;
