+++ src/main/java/com/jajaja/domain/team/service/TeamCommonService.java
@@ -3,7 +3,7 @@ package com.jajaja.domain.team.service;
 import com.jajaja.domain.team.entity.Team;
 import com.jajaja.domain.team.entity.TeamMember;
 import com.jajaja.domain.team.entity.enums.TeamStatus;
-import com.jajaja.domain.user.entity.Member;
+import com.jajaja.domain.member.entity.Member;
 import com.jajaja.global.apiPayload.code.status.ErrorStatus;
 import com.jajaja.global.apiPayload.exception.custom.BadRequestException;
 import lombok.RequiredArgsConstructor;
