--- src/main/java/com/jajaja/domain/user/entity/enums/OauthType.java
@@ -1,5 +1,0 @@
-package com.jajaja.domain.user.entity.enums;
-
-public enum OauthType {
-    KAKAO, GOOGLE
-}
